+++ front/src/components/Layout/GeneralLayout.tsx
@@ -1,19 +1,44 @@
-import { Outlet, useOutletContext } from "react-router-dom";
+import { Outlet, useNavigate, useOutletContext } from "react-router-dom";
 import Navbar from "../Navbar";
 import { NavbarContent } from "../../router";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import type { User } from "../../types";
+import { getCurrentUserInfo } from "../../api/login";
 
 type ContextType = {
   userProfile: { user_id: number; username: string } | null;
   setUserProfile: (user: { user_id: number; username: string }) => void;
 };
 
-export default function GeneralLayout() {
+export default function GeneralLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
   const [userProfile, setUserProfile] = useState<{
     user_id: number;
     username: string;
   } | null>();
+
+  const navigate = useNavigate();
+
+  const fetchUserProfile = async () => {
+    const userProfileResponse = await getCurrentUserInfo();
+
+    if (userProfileResponse === null) {
+      return (
+        confirm("로그인이 필요합니다. 로그인 하시겠습니까?") &&
+        navigate("/account/login")
+      );
+    }
+
+    setUserProfile(userProfileResponse);
+  };
+
+  useEffect(() => {
+    fetchUserProfile();
+  }, [children]);
+
   const onSetUser = (args) => {
     setUserProfile(args);
   };
@@ -21,7 +46,7 @@ export default function GeneralLayout() {
   return (
     <>
       <Navbar NavbarContent={NavbarContent} userProfile={userProfile} />
-      <Outlet context={[onSetUser]} />
+      <Outlet context={[userProfile, onSetUser]} />
     </>
   );
 }
