+++ front/src/components/Layout/Root.tsx
@@ -9,28 +9,14 @@ interface RootProps {
 }
 
 const Root = ({ children }: RootProps) => {
-  const navigate = useNavigate();
   // const { setUserProfile } = useUser();
-  const [onSetUser] = useOutletContext<{ user_id: number; username: string }>();
-  console.log(onSetUser);
-
-  const fetchUserProfile = async () => {
-    const userProfileResponse = await getCurrentUserInfo();
-
-    if (userProfileResponse === null) {
-      return (
-        confirm("로그인이 필요합니다. 로그인 하시겠습니까?") &&
-        navigate("/account/login")
-      );
-    }
-
-    onSetUser(userProfileResponse);
-  };
-
-  useEffect(() => {
-    // console.log("권한이 필요한 페이지가 바뀌었습니다.");
-    fetchUserProfile();
-  }, [children]);
+  const [userProfile, onSetUser] = useOutletContext<{
+    user_id: number;
+    username: string;
+  }>();
+  // console.log(onSetUser);
+  // console.log(userProfile);
+  console.log(useOutletContext());
 
   return (
     <div>
